+++ mobile/android/android-components/components/feature/sitepermissions/src/androidTest/java/mozilla/components/feature/sitepermissions/db/SitePermissionsDaoTest.kt
@@ -9,6 +9,7 @@ import androidx.core.net.toUri
 import androidx.room.Room
 import androidx.test.core.app.ApplicationProvider
 import mozilla.components.feature.sitepermissions.SitePermissions
+import mozilla.components.feature.sitepermissions.SitePermissions.AutoplayStatus
 import mozilla.components.feature.sitepermissions.SitePermissions.Status.ALLOWED
 import mozilla.components.feature.sitepermissions.SitePermissions.Status.BLOCKED
 import org.junit.After
@@ -45,6 +46,8 @@ class SitePermissionsDaoTest {
 
         assertEquals(origin, siteFromDb.origin)
         assertEquals(BLOCKED, siteFromDb.camera)
+        assertEquals(AutoplayStatus.BLOCKED, siteFromDb.autoplayAudible)
+        assertEquals(AutoplayStatus.ALLOWED, siteFromDb.autoplayInaudible)
     }
 
     @Test
@@ -70,12 +73,20 @@ class SitePermissionsDaoTest {
         var siteFromDb = dao.getSitePermissionsBy(origin)!!.toSitePermission()
 
         assertEquals(BLOCKED, siteFromDb.camera)
+        assertEquals(AutoplayStatus.BLOCKED, siteFromDb.autoplayAudible)
+        assertEquals(AutoplayStatus.ALLOWED, siteFromDb.autoplayInaudible)
 
-        dao.update(siteFromDb.copy(camera = ALLOWED).toSitePermissionsEntity())
+        dao.update(siteFromDb.copy(
+            camera = ALLOWED,
+            autoplayInaudible = AutoplayStatus.ALLOWED,
+            autoplayAudible = AutoplayStatus.ALLOWED
+        ).toSitePermissionsEntity())
 
         siteFromDb = dao.getSitePermissionsBy(origin)!!.toSitePermission()
 
         assertEquals(ALLOWED, siteFromDb.camera)
+        assertEquals(AutoplayStatus.ALLOWED, siteFromDb.autoplayAudible)
+        assertEquals(AutoplayStatus.ALLOWED, siteFromDb.autoplayInaudible)
 
         dao.deleteSitePermissions(siteFromDb.toSitePermissionsEntity())